pyproject.toml
[project]
name = "claimcoin-autoclaim"
version = "0.1.0"
description = "HTTP-first ClaimCoin faucet autoclaim runner"
requires-python = ">=3.11"
dependencies = [
  "Pillow>=10.0.0",
  "requests>=2.31.0",
  "PyYAML>=6.0.0",
]

[project.scripts]
claimcoin-auto = "claimcoin_autoclaim.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
